Renaming to ComputeThickness
