+++ Libs/Mesh/CMakeLists.txt
@@ -9,7 +9,7 @@ set(Mesh_sources
  meshFIM.cpp
  MeshUtils.cpp
  MeshWarper.cpp
- MeshAssignThickness.cpp
+ MeshComputeThickness.cpp
  )
 
 set(Mesh_headers
